+++ src/lib.rs
@@ -80,9 +80,9 @@ mod walk;
 pub use crate::de::{Color, Frequency, RegistriesProtocol, VersionControlSoftware, When};
 pub use crate::{
     easy::{
-        BuildConfig, Config, DocConfig, EnvConfigValue, Flags, FutureIncompatReportConfig,
-        NetConfig, PathAndArgs, RegistriesConfigValue, RegistryConfig, StringList, TargetConfig,
-        TermConfig, TermProgressConfig,
+        BuildConfig, CargoNewConfig, Config, DocConfig, EnvConfigValue, Flags,
+        FutureIncompatReportConfig, HttpConfig, NetConfig, PathAndArgs, RegistriesConfigValue,
+        RegistryConfig, StringList, TargetConfig, TermConfig, TermProgressConfig,
     },
     error::Error,
     resolve::{CargoVersion, ResolveOptions, RustcVersion, TargetTriple, TargetTripleRef},