+++ src/resolve.rs
@@ -144,6 +144,7 @@ impl ResolveOptions {
 }
 
 #[doc(hidden)] // Not public API.
+#[allow(unknown_lints, unnameable_types)] // Not public API. unnameable_types is available on Rust 1.79+
 #[derive(Debug, Clone)]
 #[must_use]
 pub struct ResolveContext {